+++ src/index.js
@@ -1,5 +1,5 @@
 import { ApolloServer, PubSub } from "apollo-server";
-import modules from "./modules";
+import schema from "./schema";
 import plugins from "./plugins";
 import database from "./database";
 
@@ -8,7 +8,7 @@ const pubsub = new PubSub();
 // Anything shared between resolvers are put into context
 const server = new ApolloServer({
   plugins,
-  schema: modules,
+  schema,
   context: ({ req }) => ({
     ...req,
     database,
