+++ src/schema/typeDefs/event/queries.graphql
